# core-banking-payments-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-baking-payments
  description: Payments Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/payments
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/payment-beneficiaries/{paymentBeneficiaryId}/favorite:
    put:
      tags:
        - Payment Beneficiaries
      summary: Mark Beneficiary as Favorite
      description: Marks a beneficiary as favorite or removes favorite status.
      operationId: markAsFavorite
      parameters:
        - name: paymentBeneficiaryId
          in: path
          description: Payment Beneficiary ID
          required: true
          schema:
            type: string
            format: uuid
        - name: isFavorite
          in: query
          description: Is Favorite
          required: true
          schema:
            type: boolean
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully updated favorite status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        '404':
          description: Beneficiary not found
  /api/v1/payment-beneficiaries/{paymentBeneficiaryId}:
    get:
      tags:
        - Payment Beneficiaries
      summary: Get Beneficiary by ID
      description: Retrieves a specific beneficiary by its ID.
      operationId: getBeneficiaryById
      parameters:
        - name: paymentBeneficiaryId
          in: path
          description: Payment Beneficiary ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved beneficiary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        '404':
          description: Beneficiary not found
    put:
      tags:
        - Payment Beneficiaries
      summary: Update Beneficiary
      description: Updates a beneficiary.
      operationId: updateBeneficiary
      parameters:
        - name: paymentBeneficiaryId
          in: path
          description: Payment Beneficiary ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        required: true
      responses:
        '200':
          description: Successfully updated beneficiary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        '404':
          description: Beneficiary not found
    delete:
      tags:
        - Payment Beneficiaries
      summary: Delete Beneficiary
      description: Deletes a beneficiary.
      operationId: deleteBeneficiary
      parameters:
        - name: paymentBeneficiaryId
          in: path
          description: Payment Beneficiary ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted beneficiary
        '404':
          description: Beneficiary not found
...
  /api/v1/payment-beneficiaries:
    get:
      tags:
        - Payment Beneficiaries
      summary: Retrieve Beneficiaries
      description: Retrieves a paginated list of beneficiaries with optional filtering.
      operationId: getAllBeneficiaries
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.payerAccountId
          in: query
          description: Exact filter for payer account id
          required: false
          schema:
            type: string
        - name: filters.beneficiaryName
          in: query
          description: Exact filter for beneficiary name
          required: false
          schema:
            type: string
        - name: filters.beneficiaryType
          in: query
          description: Exact filter for beneficiary type
          required: false
          schema:
            type: string
            enum:
              - INDIVIDUAL
              - BUSINESS
              - GOVERNMENT
        - name: filters.beneficiaryAccountNumber
          in: query
          description: Exact filter for beneficiary account number
          required: false
          schema:
            type: string
        - name: filters.beneficiaryIban
          in: query
          description: Exact filter for beneficiary iban
          required: false
          schema:
            type: string
        - name: filters.beneficiaryBic
          in: query
          description: Exact filter for beneficiary bic
          required: false
          schema:
            type: string
        - name: filters.beneficiaryAddress
          in: query
          description: Exact filter for beneficiary address
          required: false
          schema:
            type: string
        - name: filters.beneficiaryCity
          in: query
          description: Exact filter for beneficiary city
          required: false
          schema:
            type: string
        - name: filters.beneficiaryPostalCode
          in: query
          description: Exact filter for beneficiary postal code
          required: false
          schema:
            type: string
        - name: filters.beneficiaryCountryCode
          in: query
          description: Exact filter for beneficiary country code
          required: false
          schema:
            type: string
        - name: filters.beneficiaryEmail
          in: query
          description: Exact filter for beneficiary email
          required: false
          schema:
            type: string
        - name: filters.beneficiaryPhone
          in: query
          description: Exact filter for beneficiary phone
          required: false
          schema:
            type: string
        - name: filters.beneficiaryBankName
          in: query
          description: Exact filter for beneficiary bank name
          required: false
          schema:
            type: string
        - name: filters.beneficiaryBankAddress
          in: query
          description: Exact filter for beneficiary bank address
          required: false
          schema:
            type: string
        - name: filters.beneficiaryBankCountryCode
          in: query
          description: Exact filter for beneficiary bank country code
          required: false
          schema:
            type: string
        - name: filters.isFavorite
          in: query
          description: Exact filter for is favorite
          required: false
          schema:
            type: boolean
        - name: filters.nickname
          in: query
          description: Exact filter for nickname
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved beneficiaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No beneficiaries found
    post:
      tags:
        - Payment Beneficiaries
      summary: Create Beneficiary
      description: Creates a new beneficiary.
      operationId: createBeneficiary
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        required: true
      responses:
        '201':
          description: Successfully created beneficiary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentBeneficiaryDTO'
        '400':
          description: Invalid input
...
components:
  schemas:
...
    PaymentBeneficiaryDTO:
      required:
        - beneficiaryName
        - beneficiaryType
        - isFavorite
        - payerAccountId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        paymentBeneficiaryId:
          type: string
          format: uuid
          readOnly: true
        payerAccountId:
          type: string
          format: uuid
        beneficiaryName:
          maxLength: 255
          minLength: 0
          type: string
        beneficiaryType:
          type: string
          enum:
            - INDIVIDUAL
            - BUSINESS
            - GOVERNMENT
        beneficiaryAccountNumber:
          maxLength: 50
          minLength: 0
          type: string
        beneficiaryIban:
          pattern: ^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$
          type: string
        beneficiaryBic:
          pattern: ^[A-Z]{6}[A-Z0-9]{2}([A-Z0-9]{3})?$
          type: string
        beneficiaryAddress:
          maxLength: 500
          minLength: 0
          type: string
        beneficiaryCity:
          maxLength: 100
          minLength: 0
          type: string
        beneficiaryPostalCode:
          pattern: ^[A-Z0-9]{3,10}$
          type: string
        beneficiaryCountryCode:
          pattern: ^[A-Z]{2}$
          type: string
        beneficiaryEmail:
          maxLength: 255
          minLength: 0
          type: string
        beneficiaryPhone:
          pattern: ^\+?[1-9]\d{1,14}$
          type: string
        beneficiaryTaxId:
          maxLength: 50
          minLength: 0
          type: string
        beneficiaryBankName:
          maxLength: 255
          minLength: 0
          type: string
        beneficiaryBankAddress:
          maxLength: 500
          minLength: 0
          type: string
        beneficiaryBankCountryCode:
          pattern: ^[A-Z]{2}$
          type: string
        isFavorite:
          type: boolean
        nickname:
          maxLength: 100
          minLength: 0
          type: string
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...